% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{mcmc function}
\usage{
mcmc(ct, u1, u2, X, nthin = 1)
}
\arguments{
\item{ct}{survival response, a \eqn{n*2} matrix with first column as response and second column as right censored indicator,
1 is event time and 0 is right censored.}

\item{u2}{Matix of predictors from the first platform, dimension \eqn{n*q_2}}

\item{X}{Matrix of covariates, dimension \eqn{n*p}.
'@param u1 Matix of predictors from the first platform, dimension \eqn{n*q_1}}

\item{nthin}{thinning parameter. Default is 1 (no thinning)}
}
\description{
MCMC routine for the strucatural equation model
}
